[#1] more initial setup (#4)
+ add windows 10 hardening to the host
+ add a simple direct2d backend for windows
+ tweak the manifest to only specify windows 10 (win 7 and 8 are still supported)

+++ src/platform/windows/win_ime.cpp
@@ -153,7 +153,7 @@ HRESULT __stdcall text_input_manager_old::GetWnd(TsViewCookie vcView, HWND* phwn
 text_input_manager::text_input_manager()
     : m_old(*this)
 {
-    ::ImmDisableIME(-1);
+    ::ImmDisableIME(static_cast<DWORD>(-1));
 
     if (auto const hr = ::CoCreateInstance(
               CLSID_TF_ThreadMgr
